conformance/buffimgs/cl_commandQueue_enqueueReadImage.html

<!--
 Copyright (C) 2013 Samsung Electronics Corporation. All rights reserved.

 Redistribution and use in source and binary forms, with or without
 modification, are permitted provided the following condition
 are met:

 1.  Redistributions of source code must retain the above copyright
     notice, this list of conditions and the following disclaimer.

 2.  Redistributions in binary form must reproduce the above copyright
     notice, this list of conditions and the following disclaimer in the
     documentation and/or other materials provided with the distribution.

 THIS SOFTWARE IS PROVIDED BY SAMSUNG ELECTRONICS CORPORATION AND ITS
 CONTRIBUTORS "AS IS", AND ANY EXPRESS OR IMPLIED WARRANTIES, INCLUDING
 BUT NOT LIMITED TO, THE IMPLIED WARRANTIES OF MERCHANTABILITY AND FITNESS
 FOR A PARTICULAR PURPOSE ARE DISCLAIMED. IN NO EVENT SHALL SAMSUNG
 ELECTRONICS CORPORATION OR ITS CONTRIBUTORS BE LIABLE FOR ANY DIRECT,
 INDIRECT, INCIDENTAL, SPECIAL, EXEMPLARY, OR CONSEQUENTIAL DAMAGES(INCLUDING
 BUT NOT LIMITED TO, PROCUREMENT OF SUBSTITUTE GOODS OR SERVICES; LOSS OF USE,
 DATA, OR PROFITS, OR BUSINESS INTERRUPTION), HOWEVER CAUSED AND ON ANY THEORY
 OF LIABILITY, WHETHER IN CONTRACT, STRICT LIABILITY, OR TORT(INCLUDING
 NEGLIGENCE OR OTHERWISE ARISING IN ANY WAY OUT OF THE USE OF THIS SOFTWARE,
 EVEN IF ADVISED OF THE POSSIBILITY OF SUCH DAMAGE.
 -->
<html>
<head>
<meta charset="utf-8">
    <title>WebCLCommandQueue enqueueReadImage conformance test.</title>
    <link rel="stylesheet" href="../../resources/js-test-style.css"/>
    <script src="../../resources/js-test-pre.js"></script>
    <script src="../../resources/cl_utils.js"></script>
</head>
<body>
<div id="description"></div>
<div id="console"></div>
<script>

try {
    debug("WebCLCommandQueue :: enqueueReadImage() test case. \n ");
    debug("(WebCLImage image, CLboolean blockingRead, CLuint[2] origin, CLuint[2] region, CLuint hostRowPitch, ArrayBuffer hostPtr, optional WebCLEvent[]? eventWaitList, optional WebCLEvent? event)");

    var wtu = WebCLTestUtils;
    var webCLContext = wtu.createContext();
    var webCLCommandQueue = wtu.createCommandQueue(webCLContext);
    var origin = new Int32Array([12, 25]);
    var region = new Int32Array([10, 20]);
    var data = new Float32Array(DATA_SIZE);
    var imageWidth = 100;
    var imageHeight = 100;

    // MEM_READ_ONLY images.
    var webCLImageDescriptorObject = wtu.getSupportedImageFormats(webCLContext, webcl.MEM_READ_ONLY, imageWidth, imageHeight);
    var webCLImageRead = webCLContext.createImage(webcl.MEM_READ_ONLY, webCLImageDescriptorObject);
    shouldBeTrue("wtu.enqueueReadImage(webCLCommandQueue, webCLImageRead, true, origin, region, 0, data);");
    shouldBeTrue("wtu.enqueueReadImage(webCLCommandQueue, webCLImageRead, false, origin, region, 0, data);");

    // MEM_WRITE_ONLY images.
    webCLImageDescriptorObject = wtu.getSupportedImageFormats(webCLContext, webcl.MEM_WRITE_ONLY, imageWidth, imageHeight);
    var webCLImageWrite = webCLContext.createImage(webcl.MEM_WRITE_ONLY, webCLImageDescriptorObject);
    shouldBeTrue("wtu.enqueueReadImage(webCLCommandQueue, webCLImageWrite, true, origin, region, 0, data);");
    shouldBeTrue("wtu.enqueueReadImage(webCLCommandQueue, webCLImageWrite, false, origin, region, 0, data);");

    // MEM_READ_WRITE images.
    webCLImageDescriptorObject = wtu.getSupportedImageFormats(webCLContext, webcl.MEM_READ_WRITE, imageWidth, imageHeight);
    var webCLImageReadWrite = webCLContext.createImage(webcl.MEM_READ_WRITE, webCLImageDescriptorObject);
    shouldBeTrue("wtu.enqueueReadImage(webCLCommandQueue, webCLImageReadWrite, true, origin, region, 0, data);");
    shouldBeTrue("wtu.enqueueReadImage(webCLCommandQueue, webCLImageReadWrite, false, origin, region, 0, data);");

    // FIXME :: Need to add test cases passing event parameters.

    // Testing for values of Image Offsets.
    var originInvalid = new Int32Array([imageWidth + 1, imageHeight + 1]);
    shouldThrowExceptionName("wtu.enqueueReadImage(webCLCommandQueue, webCLImageReadWrite, true, originInvalid, region, 0, data);", "INVALID_VALUE");

    // Testing for values of rectangular region.
    var regionInvalid = new Int32Array([imageWidth + 1, imageHeight + 1]);
    shouldThrowExceptionName("wtu.enqueueReadImage(webCLCommandQueue, webCLImageReadWrite, true, origin, regionInvalid, 0, data);", "INVALID_VALUE");

    // Testing for values of hostPointer.
    shouldThrowExceptionName("wtu.enqueueReadImage(webCLCommandQueue, webCLImageRead, true, origin, region, -1, data);", "INVALID_VALUE");

    // Testing invalid objects as params.
    shouldThrowExceptionName("wtu.enqueueReadImage(webCLCommandQueue, invalid_object, true, origin, region, 0, data);", "INVALID_MEM_OBJECT");
    shouldThrowExceptionName("wtu.enqueueReadImage(webCLCommandQueue, webCLImageReadWrite, invalid_object, origin, region, 0, data);", "INVALID_VALUE");
    shouldThrowExceptionName("wtu.enqueueReadImage(webCLCommandQueue, webCLImageReadWrite, true, invalid_object, region, 0, data);", "INVALID_VALUE");
    shouldThrowExceptionName("wtu.enqueueReadImage(webCLCommandQueue, webCLImageReadWrite, true, origin, invalid_object, 0, data);", "INVALID_VALUE");
    shouldThrowExceptionName("wtu.enqueueReadImage(webCLCommandQueue, webCLImageReadWrite, true, origin, region, invalid_object, data);", "INVALID_VALUE");
    shouldThrowExceptionName("wtu.enqueueReadImage(webCLCommandQueue, webCLImageReadWrite, true, origin, region, 0, invalid_object);", "INVALID_VALUE");

    // Testing for EXCEPTIONS
    var webCLContext2 = wtu.createContext();
    var webCLImage2 = webCLContext2.createImage(webcl.MEM_READ_WRITE, webCLImageDescriptorObject);
    shouldThrowExceptionName("wtu.enqueueReadImage(webCLCommandQueue, webCLImage2, true, origin, region, 0, data);", "INVALID_CONTEXT");

    var webCLCommandqueue2 = wtu.createCommandQueue(webCLContext2);
    shouldThrowExceptionName("wtu.enqueueReadImage(webCLCommandqueue2, webCLImageReadWrite, true, origin, region, 0, data);", "INVALID_CONTEXT");

    var hostRowPitchInvalid = -1;
    shouldThrowExceptionName("wtu.enqueueReadImage(webCLCommandQueue, webCLImageReadWrite, true, origin, region, hostRowPitchInvalid, data);", "INVALID_CONTEXT");

    // FIXME :: Need to add test cases passing event parameters for INVALID_EVENT_WAIT_LIST and EXEC_STATUS_ERROR_FOR_EVENTS_IN_WAIT_LIST
    // FIXME Need to add test case for INVALID_IMAGE_SIZE.

} catch (e) {
    testFailed("WebCLCommandQueue :: enqueueReadImage() throws an exception : " + e.message);
}
notifyFinishedToHarness();
successfullyParsed = true;
</script>
<script src="../../resources/js-test-post.js"></script>
</body>
</html>
